wip(cucumber): partially build creatives_data_table.feature
It looks like chai-as-promised is throwing exceptions instead of
rejecting promises. This normally isn't a huge problem, but in this
case, we're wrapped in many layers of closures and the test hangs
on failure. If you take the q.allSettled() thingy out, it fails
without hanging.  But...  WTF?  WHY!?!  *dies*