fix(angular-loading-bar): add bug fix and css
I created a fix for the loading bar that addresses a bug
The bar no longer breaks when the ui-view is attached to the
body.  I have a pull request into the origin project on github.
Until that's resolved, I updated bower to pull from my fork.

I updated index.html to include a place for bower css to be
injected.  It's correctly pulling in the loading bar css again.