fix(jshint): remove unsupported indent config
The newer versions of jshint do not support indent
checking - which is totally lame.  I took out the
entry to prevent some consusion.
We need to decide on an indentation practice and
stick with it.  There's a lot of support for 2-space,
so it seems like we should just do that.