chore(sticky_nav) implement nav bar sticky position
Polyfills out there did not help, as some of the elements
calculations were made during the initialization phase of
the polyfill. This is a port for angularjs gathering ideas
from various sources on how to implement sticky position.
There are a couple of known issues: does not re-calculate
on screen resize and it keeps at viewpane's 0x0 on horizontal
scroll.

Tests, anyone? Be my guest! :)